+++ frontend/host-app/src/components/Questions/CompleteQuestionCard.js
@@ -14,7 +14,7 @@ import QuestionUserName from "./QuestionUserName";
 import useStyles from "./useStyles";
 import QuestionMenu from "./QuestionMenu";
 import ThumbUpButton from "./ThumbUpButton";
-import Replies from "./Replies";
+import ReplyList from "../Reply/ReplyList.js";
 import RestoreQuestionIconButton from "./RestoreQuestionIconButton.js";
 import QuestionCardInnerDivider from "./QuestionCardInnerDivider.js";
 
@@ -54,7 +54,7 @@ function CompleteQuestionCard(props) {
 					replyOpenHandler={setOpenReplies}
 					replyOpenStatus={openReplies}
 				/>
-				{openReplies && <Replies replies={props.replies} />}
+				{openReplies && <ReplyList replies={props.replies} />}
 			</CardContent>
 		</Card>
 	);
