+++ frontend/host-app/src/components/Reply/Reply.js
@@ -1,0 +1,33 @@
+import React from "react";
+import {Icon, Typography} from "@material-ui/core";
+import useStyles from "../Questions/useStyles.js";
+import {QuestionMeta, ReplyBody, ReplyContainer, ReplyInfo} from "../Questions/QuestionStyle.js";
+import QuestionUserName from "../Questions/QuestionUserName.js";
+import QuestionDate from "../Questions/QuestionDate.js";
+
+function Reply(props) {
+	const classes = useStyles();
+
+	return (
+		<>
+			<QuestionMeta>
+				<ReplyInfo>
+					<Icon className={classes.replyIcon}>
+						subdirectory_arrow_right
+					</Icon>
+					<QuestionUserName {...props} />
+					<ReplyContainer>
+						<QuestionDate {...props} />
+					</ReplyContainer>
+				</ReplyInfo>
+			</QuestionMeta>
+			<ReplyBody>
+				<Typography color={"textPrimary"} variant={"body2"}>
+					{props.content}
+				</Typography>
+			</ReplyBody>
+		</>
+	);
+}
+
+export default Reply;
